fix(authority): restore recovery evidence membership
